moved DiffusionDecoder to decoders/diffusion_decoder.py
