+++ decoders/diffusion_decoder.py
@@ -1,0 +1,48 @@
+import torch
+from torch import optim, nn
+from torch.nn import functional as F
+from encoders.encoders import RAVEEncoder, ResConvBlock
+from diffusion.model import SkipBlock, FourierFeatures, expand_to_planes
+
+class DiffusionDecoder(nn.Module):
+    def __init__(self, latent_dim, io_channels, depth=16):
+        super().__init__()
+        max_depth = 16
+        depth = min(depth, max_depth)
+        c_mults = [256, 256, 512, 512] + [512] * 12
+        c_mults = c_mults[:depth]
+
+        self.io_channels = io_channels
+        self.timestep_embed = FourierFeatures(1, 16)
+        block = nn.Identity()
+        for i in range(depth, 0, -1):
+            c = c_mults[i - 1]
+            if i > 1:
+                c_prev = c_mults[i - 2]
+                block = SkipBlock(
+                    nn.AvgPool1d(2),
+                    ResConvBlock(c_prev, c, c),
+                    ResConvBlock(c, c, c),
+                    ResConvBlock(c, c, c),
+                    block,
+                    ResConvBlock(c * 2 if i != depth else c, c, c),
+                    ResConvBlock(c, c, c),
+                    ResConvBlock(c, c, c_prev),
+                    nn.Upsample(scale_factor=2, mode='linear', align_corners=False),
+                )
+            else:
+                block = nn.Sequential(
+                    ResConvBlock(io_channels + 16 + latent_dim, c, c),
+                    ResConvBlock(c, c, c),
+                    ResConvBlock(c, c, c),
+                    block,
+                    ResConvBlock(c * 2, c, c),
+                    ResConvBlock(c, c, c),
+                    ResConvBlock(c, c, io_channels, is_last=True),
+                )
+        self.net = block
+
+    def forward(self, input, t, quantized):
+        timestep_embed = expand_to_planes(self.timestep_embed(t[:, None]), input.shape)
+        quantized = F.interpolate(quantized, (input.shape[2], ), mode='linear', align_corners=False)
+        return self.net(torch.cat([input, timestep_embed, quantized], dim=1))
